Create RegenAuthoritySystem for handling regen
